requirements.txt
fastapi
uvicorn
streamlit
groq
python-dotenv
requests